+++ ieee80211_util.h
@@ -25,7 +25,7 @@ u8*
 ieee80211_get_bssid(struct ieee80211_hdr *hdr, size_t len);
 
 void
-ieee802_11_parse_elems(unsigned char *start, int len, struct packet_info *pkt);
+ieee802_11_parse_elems(unsigned char *start, size_t len, struct packet_info *pkt);
 
 int
 ieee80211_frame_duration(int phymode, size_t len,
